--- library/src/main/java/jp/co/soramitsu/feature/maps/impl/data/PlacesRepository.kt
@@ -1,7 +1,0 @@
-package jp.co.soramitsu.feature.maps.impl.data
-
-import jp.co.soramitsu.feature.maps.impl.model.Place
-
-interface PlacesRepository {
-    fun getAllPlaces(): List<Place>
-}
