--- library/src/main/java/jp/co/soramitsu/feature/maps/impl/model/Author.kt
@@ -1,6 +1,0 @@
-package jp.co.soramitsu.feature.maps.impl.model
-
-class Author(
-    val name: String,
-    val avatarUrl: String = ""
-)
