--- library/src/main/java/jp/co/soramitsu/feature/maps/impl/model/Category.kt
@@ -1,12 +1,0 @@
-package jp.co.soramitsu.feature.maps.impl.model
-
-enum class Category {
-    BANK,
-    SERVICES,
-    FOOD,
-    SUPERMARKETS,
-    PHARMACY,
-    ENTERTAINMENT,
-    EDUCATION,
-    OTHER
-}
