--- library/src/main/java/jp/co/soramitsu/feature/maps/impl/model/Position.kt
@@ -1,6 +1,0 @@
-package jp.co.soramitsu.feature.maps.impl.model
-
-class Position(
-    val latitude: Double,
-    val longitude: Double
-)
