--- library/src/main/java/jp/co/soramitsu/feature/maps/impl/model/Review.kt
@@ -1,8 +1,0 @@
-package jp.co.soramitsu.feature.maps.impl.model
-
-data class Review(
-    val author: Author,
-    val rating: Float,
-    val date: Long,
-    val text: String
-)
