--- library/src/main/java/jp/co/soramitsu/feature/maps/impl/model/Time.kt
@@ -1,6 +1,0 @@
-package jp.co.soramitsu.feature.maps.impl.model
-
-data class Time(
-    val hour: Int,
-    val minute: Int
-)
