--- library/src/main/java/jp/co/soramitsu/feature/maps/impl/presentation/categories/CategoriesAdapter.kt
@@ -1,87 +1,0 @@
-package jp.co.soramitsu.feature.maps.impl.presentation.categories
-
-import android.content.Context
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.appcompat.widget.AppCompatTextView
-import androidx.recyclerview.widget.DiffUtil
-import androidx.recyclerview.widget.RecyclerView
-import jp.co.soramitsu.feature.maps.impl.ext.asLocalizedString
-import jp.co.soramitsu.feature.maps.impl.ext.getResourceIdForAttr
-import jp.co.soramitsu.feature.maps.impl.model.Category
-import jp.co.soramitsu.feature_map_impl.R
-
-class CategoriesAdapter : RecyclerView.Adapter<CategoryViewHolder>() {
-
-    private val items = mutableListOf<CategoryListItem>()
-
-    var onCategoryClick: (Category) -> Unit = {}
-
-    fun update(categories: List<CategoryListItem>) {
-        val result = DiffUtil.calculateDiff(CategoryDiffUtilCallback(items, categories))
-        items.clear()
-        items.addAll(categories)
-        result.dispatchUpdatesTo(this)
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CategoryViewHolder {
-        val inflater = LayoutInflater.from(parent.context)
-        val view = inflater.inflate(R.layout.category_list_item, parent, false)
-        val viewHolder = CategoryViewHolder(view)
-        view.setOnClickListener {
-            onCategoryClick.invoke(items[viewHolder.adapterPosition].category)
-        }
-        return viewHolder
-    }
-
-    override fun getItemCount(): Int = items.size
-
-    override fun onBindViewHolder(holder: CategoryViewHolder, position: Int) {
-        holder.bind(items[position])
-    }
-}
-
-class CategoryDiffUtilCallback(
-    private val beforeList: List<CategoryListItem>,
-    private val afterList: List<CategoryListItem>
-) : DiffUtil.Callback() {
-    override fun areItemsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
-        return beforeList[oldItemPosition].category.name == afterList[newItemPosition].category.name
-    }
-
-    override fun getOldListSize(): Int = beforeList.size
-
-    override fun getNewListSize(): Int = afterList.size
-
-    override fun areContentsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
-        return beforeList[oldItemPosition] == afterList[newItemPosition]
-    }
-}
-
-class CategoryViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-
-    private val categoryTextView =
-        itemView.findViewById<AppCompatTextView>(R.id.categoryNameTextView)
-
-    fun bind(categoryListItem: CategoryListItem) {
-        categoryTextView.setText(categoryListItem.category.asLocalizedString())
-        categoryTextView.isSelected = categoryListItem.selected
-        val checkIconId = if (categoryListItem.selected) R.drawable.ic_check_24dp else 0
-        categoryTextView.setCompoundDrawablesRelativeWithIntrinsicBounds(
-            iconForCategory(categoryTextView.context, categoryListItem.category), 0, checkIconId, 0
-        )
-    }
-
-    private fun iconForCategory(context: Context, category: Category): Int = when (category) {
-        Category.BANK -> context.getResourceIdForAttr(R.attr.soramitsuMapCategoryIconDeposit)
-        Category.FOOD -> context.getResourceIdForAttr(R.attr.soramitsuMapCategoryIconRestaurant)
-        Category.SERVICES -> context.getResourceIdForAttr(R.attr.soramitsuMapCategoryIconServices)
-        Category.SUPERMARKETS -> context.getResourceIdForAttr(R.attr.soramitsuMapCategoryIconSupermarket)
-        Category.PHARMACY -> context.getResourceIdForAttr(R.attr.soramitsuMapCategoryIconPharmacy)
-        Category.ENTERTAINMENT -> context.getResourceIdForAttr(R.attr.soramitsuMapCategoryIconEntertainment)
-        Category.EDUCATION -> context.getResourceIdForAttr(R.attr.soramitsuMapCategoryIconEducation)
-        Category.OTHER -> context.getResourceIdForAttr(R.attr.soramitsuMapCategoryIconOther)
-    }
-
-}
