--- library/src/main/java/jp/co/soramitsu/feature/maps/impl/presentation/categories/CategoryListItem.kt
@@ -1,8 +1,0 @@
-package jp.co.soramitsu.feature.maps.impl.presentation.categories
-
-import jp.co.soramitsu.feature.maps.impl.model.Category
-
-data class CategoryListItem(
-    val category: Category,
-    val selected: Boolean = false
-)
