--- library/src/main/java/jp/co/soramitsu/feature/maps/impl/presentation/lifycycle/SingleLiveEvent.kt
@@ -1,51 +1,0 @@
-package jp.co.soramitsu.feature.maps.impl.presentation.lifycycle
-
-import androidx.annotation.MainThread
-import androidx.annotation.Nullable
-import androidx.lifecycle.LifecycleOwner
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.Observer
-import java.util.concurrent.atomic.AtomicBoolean
-
-/**
- * A lifecycle-aware observable that sends only new updates after subscription, used for events like
- * navigation and Snackbar messages.
- * <p>
- * This avoids a common problem with events: on configuration change (like rotation) an update
- * can be emitted if the observer is active. This LiveData only calls the observable if there's an
- * explicit call to setValue() or call().
- * <p>
- * Note that only one observer is going to be notified of changes.
- */
-class SingleLiveEvent<T> : MutableLiveData<T> {
-
-    private val mPending = AtomicBoolean(false)
-
-    constructor(value: T) : super(value)
-    constructor() : super()
-
-    @MainThread
-    override fun observe(owner: LifecycleOwner, observer: Observer<in T>) {
-
-        // Observe the internal MutableLiveData
-        super.observe(owner, Observer<T> { t ->
-            if (mPending.compareAndSet(true, false)) {
-                observer.onChanged(t)
-            }
-        })
-    }
-
-    @MainThread
-    override fun setValue(@Nullable t: T?) {
-        mPending.set(true)
-        super.setValue(t)
-    }
-
-    /**
-     * Used for cases where T is Void, to make calls cleaner.
-     */
-    @MainThread
-    fun call() {
-        value = null
-    }
-}
