+++ library/src/main/java/jp/co/soramitsu/map/data/Banks.kt
@@ -1,0 +1,1064 @@
+package jp.co.soramitsu.map.data
+
+import jp.co.soramitsu.map.model.Category
+import jp.co.soramitsu.map.model.Place
+import jp.co.soramitsu.map.model.Position
+
+/**
+ * All banks and branches on geo map
+ */
+val Places.banks
+    get() = listOf(
+        // region banks
+        Place(
+            name = "Sathapana Bank. Main Branch",
+            category = Category.BANK,
+            position = Position(11.563457, 104.926148),
+            phone = "(+855) 96 265 1111",
+            address = "#83, Norodom Blvd., Sangkat Phsar Thmey3, Khan Daun Penh, Phnom Penh."
+        ),
+        Place(
+            name = "Sathapana Bank. Chamkamorn Branch",
+            category = Category.BANK,
+            position = Position(11.530912, 104.926022),
+            phone = "(+855) 96 904 7777",
+            address = "#831c, Monivong Blvd., Sangkat Phsar Daeum Thkov, Khan Chamkar Mon, Phnom Penh Capital."
+        ),
+        Place(
+            name = "Sathapana Bank. Phnom Penh",
+            category = Category.BANK,
+            position = Position(11.565651, 104.911473),
+            phone = "(+855) 96 916 7777",
+            address = "#D1-D2, St. 169, Sangkat Veal Vong, Khan Prampir Meakkakra, Phnom Penh Capital."
+        ),
+        Place(
+            name = "Sathapana Bank. Pou Senchey Branch",
+            category = Category.BANK,
+            position = Position(11.564485, 104.860070),
+            phone = "(+855) 96 805 7777",
+            address = "# 325B, Sahakpenrossei Blvd., Group 6, Paprak Khang Cheung Village, Kakap Sangkat, Pur Senchey, Phnom Penh Capital."
+        ),
+        Place(
+            name = "Sathapana Bank. Russey Keo Branch",
+            category = Category.BANK,
+            position = Position(11.615973, 104.915018),
+            phone = "(+855) 96 840 7777",
+            address = "National Road No.5, Kraol Kou Village, Kilomaetr Lekh Prammuoy Sangkat, Russey Keo Khan, Phnom Penh Capital."
+        ),
+        Place(
+            name = "Sathapana Bank. Phsar Deumkor Branch",
+            category = Category.BANK,
+            position = Position(11.551061, 104.904263),
+            phone = "(+855) 96 841 7777",
+            address = "# 284 BC, St. MaoTseToung, Phum 4, Tom Nup Teok Sangkat, Chamkar Mon Khan, Phnom Penh Capital."
+        ),
+        Place(
+            name = "Sathapana Bank. Stueng Meanchey Branch",
+            category = Category.BANK,
+            position = Position(11.540529, 104.889948),
+            phone = "(+855) 96 848 7777",
+            address = "#206, St. Monireth, Trea Village, Stueng Mean chey Sangkat, Mean chey Khan, Phnom Penh Capital."
+        ),
+        Place(
+            name = "Sathapana Bank. Chamkar Dong Branch",
+            category = Category.BANK,
+            position = Position(11.515785, 104.889823),
+            phone = "(+855) 96 776 1111",
+            address = "# A12, St. 217, Sambour Village, Dangkao Sangkat, Dangkao Khan, Phnom Penh Capital."
+        ),
+        Place(
+            name = "Sathapana Bank. Tuol Kork Branch",
+            category = Category.BANK,
+            position = Position(11.583840, 104.900203),
+            phone = "(+855) 96 851 7777",
+            address = "#24DEFG, Street 273, Phum 14, Sangkat Boeng Kak Ti Muoy, Khan Tuol Kouk, Phnom Penh."
+        ),
+        Place(
+            name = "Sathapana Bank. Prampi Makara Branch",
+            category = Category.BANK,
+            position = Position(11.562538, 104.919508),
+            phone = "(+855) 96 803 7777",
+            address = "# 2E0, 2E1, 2E2, 2E3, St. 182+93, Boeung Proluet Sangkat, Prampir Meakkakra, Phnom Penh Capital"
+        ),
+        Place(
+            name = "Sathapana Bank. Chom Chao Branch",
+            category = Category.BANK,
+            position = Position(11.537487, 104.822875),
+            phone = "(+855) 96 912 7777",
+            address = "#A6-7, National Road No.4, Thnal Bambaek Village, Chaom Chau Sangkat, Pur Sen chey Khan, Phnom Penh Capital."
+        ),
+        Place(
+            name = "Sathapana Bank. Phsar Hengly Branch",
+            category = Category.BANK,
+            position = Position(11.554403, 104.887207),
+            phone = "(+855) 96 765 7777",
+            address = "#1+ 2AEo, Street 271, Phum 11, Sangkat Teuk Lak Ti Bei , Khan Toul Kouk , Phnom Penh"
+        ),
+        Place(
+            name = "Sathapana Bank. AEON Mall Sensok Branch",
+            category = Category.BANK,
+            position = Position(11.600658, 104.885197),
+            phone = "(+855) 96 903 7777",
+            address = "2nd Floor, AEON Mall 2, Bayarb Village, Sangkat Phnom Penh Thmei, Khan Sen Sok, Phnom Penh."
+        ),
+        Place(
+            name = "Sathapana Bank. Chbar Ampov Branch",
+            category = Category.BANK,
+            position = Position(11.531326, 104.948885),
+            phone = "(+855) 96 349 2222",
+            address = "#44, National Road 1, Boeng Chhuk Village, Sangkat Nirouth, Khan Chbar Ampov, Phnom Penh"
+        ),
+        Place(
+            name = "Sathapana Bank. Riverside Branch",
+            category = Category.BANK,
+            position = Position(11.575929, 104.926841),
+            phone = "(+855) 96 367 1111",
+            address = "#91, Preah Sisovat Boulevard, Phum 8, Sangkat Voat Phnum, Khan Doun Penh, Phnom Penh"
+        ),
+        Place(
+            name = "Sathapana Bank. Phnom Penh Thmei Branch",
+            category = Category.BANK,
+            position = Position(11.575262, 104.870532),
+            phone = "(+855) 96 359 1111",
+            address = "#583, Phnom Penh Hanoi Friendship Boulevard (1019), Phnom Penh Thmei Village, Sangkat Phnom Penh Thmei, Khan Saensokh, Phnom Penh"
+        ),
+        Place(
+            name = "Sathapana Bank. Kandal Provincial Branch",
+            category = Category.BANK,
+            position = Position(11.470985, 104.943220),
+            phone = "(+855) 96 806 7777",
+            address = "Thmei Village, Ta Khmao Sangkat, Ta Khmau Municipality, Kandal Province."
+        ),
+        Place(
+            name = "Sathapana Bank. S'ang District Branch - Svay Rolum Commune",
+            category = Category.BANK,
+            position = Position(11.455223, 104.003559),
+            phone = "(+855) 96 8150 777",
+            address = "Lekh Buon Village, Svay Rolum Commune, S’ang District, Kandal Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Kaoh Thum District Branch - Preaek Thmei Commune",
+            category = Category.BANK,
+            position = Position(11.128392, 105.056869),
+            phone = "(+855) 96 810 7777",
+            address = "St. 21, Group 9, Preaek Thmie Village, Preaek Thmie Commune, Kaoh Thum Distirct, Kandal Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Kien Svay District Branch - Kokir Commune",
+            category = Category.BANK,
+            position = Position(11.509985, 105.052140),
+            phone = "(+855) 96 808 7777",
+            address = "National Road No.1, Tuol Tnaot Village, Kokir Communce, Kien Svay District, Kandal Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Angk Snuol District Branch - Baek Chan Commune",
+            category = Category.BANK,
+            position = Position(11.509743, 104.741801),
+            phone = "(+855) 96 802 7777",
+            address = "National Road No.4, Svay Chrum Village, Baek Chan Commune, Angk Snuol District, Kandal Province"
+        ),
+        Place(
+            name = "Sathapana Bank. Krong Siem Reap Branch",
+            category = Category.BANK,
+            position = Position(13.361282, 103.874106),
+            phone = "(+855) 96 934 7777",
+            address = "National Road #6, Chongkoasou Village, Sla Kram Commune, Siem Reap Municipality, Siem Reap Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Siem Reap Provincial Branch",
+            category = Category.BANK,
+            position = Position(13.367013, 103.849665),
+            phone = "(+855) 96 811 7777",
+            address = "National Road No. 6, Ta Phul Village, Svay Dankum Sangkat, Siem Reap Municipality, Siem Reap Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Kampong Cham Provincial Branch",
+            category = Category.BANK,
+            position = Position(12.0276229, 105.452006),
+            phone = "(+855) 96 816 7777",
+            address = "National Road No. 7A, Phum Ti Prammuoy, Veal Vong Sangkat, Kampong Cham Municipality, Kampong Cham Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Kampong Thom Provincial Branch",
+            category = Category.BANK,
+            position = Position(12.741163, 104.893066),
+            phone = "(+855) 96 821 7777",
+            address = "National Road No.6, Kampong Thum Village, Kampong Roteh Sangkat, Stueng Saen Municipality, Kampong Thom Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Battambang Provincial Branch",
+            category = Category.BANK,
+            position = Position(13.106347, 103.200368),
+            phone = "(+855) 96 828 7777",
+            address = "National Road No. 5, Prek Moha Tep Village, Svay Por Sangkat, Battambang Municipality, Battambang Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Krong Battambang Branch",
+            category = Category.BANK,
+            position = Position(13.090735, 103.211848),
+            phone = "(+855) 96 581 2222",
+            address = "Rum Chek 4 Village, Rattanak Commune, Battambang City, Battambang Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Kampong Speu Provincial Branch",
+            category = Category.BANK,
+            position = Position(11.466705, 104.530011),
+            phone = "(+855) 96 830 7777",
+            address = "National Road No. 4, Borei Kammeakkar Village, Rokar Thum Sangkat, Chbar Mon Municipality, Kampong Speu Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Kampong Chhnang Provincial Branch",
+            category = Category.BANK,
+            position = Position(12.246348, 104.670762),
+            phone = "(+855) 96 825 7777",
+            address = "National Road No. 5, La Tuek Trei Village, Kampong Chhnang Sangkat, Kampong Chhnang Municipality, Kampong Chhnang Province"
+        ),
+        Place(
+            name = "Sathapana Bank. Kampot Provincial Branch",
+            category = Category.BANK,
+            position = Position(10.619274, 104.182968),
+            phone = "(+855) 96 831 7777",
+            address = "National Road No. 3, Svay Thum Village, Krang Ampil Sangkat, Kampot Municipality, Kampot Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Kep Provincial Branch",
+            category = Category.BANK,
+            position = Position(10.539524, 104.322003),
+            phone = "(+855) 96 8151 777",
+            address = "National Road No. 33, Damnak Chang’aeur Village, Prey Thum Sangkat, Kaeb Municipality, Kep Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Pursat Provincial Branch",
+            category = Category.BANK,
+            position = Position(12.539011, 103.914056),
+            phone = "(+855) 96 826 7777",
+            address = "National Road No.5, Peal Nheaek Muoy Village, Phteah Prey Sangkat, Pursat Municipality, Pursat Province"
+        ),
+        Place(
+            name = "Sathapana Bank. Banteay Meanchey Provincial Branch",
+            category = Category.BANK,
+            position = Position(13.606804, 102.961717),
+            phone = "(+855) 96 829 7777",
+            address = "#627-628, National Road No. 5, Roung Masin Village, Ou Ambel Sangkat, Serei Saophoan Municipality, Banteay Meanchey Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Krong Paoy Paet Branch",
+            category = Category.BANK,
+            position = Position(13.677579, 102.556784),
+            phone = "(+855) 96 860 7777",
+            address = "National Raod No. 5, Kbal Spean Village, Paoy Paet Sangkat, Paoy Paet Municipalityt, Banteay Meanchay Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Preah Sihanouk Provincial Branch",
+            category = Category.BANK,
+            position = Position(10.617514, 103.538466),
+            phone = "(+855) 96 838 7777",
+            address = "#72, Ekareach Street, Group 14, Phum Bei, Buon Sangkat, Preah Sihanouk Municipality, Preah Sihanouk Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Takeo Provincial Branch",
+            category = Category.BANK,
+            position = Position(10.989467, 104.777071),
+            phone = "(+855) 96 834 7777",
+            address = "# 158, St. 20, Thnalbaek Village, Roka Krao Sangkat, Doun Kaev Municipality, Takeo Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Prey Veng Provincial Branch",
+            category = Category.BANK,
+            position = Position(11.507874, 105.330026),
+            phone = "(+855) 96 846 7777",
+            address = "National Road No.11 Lekh Prampir Village, Kampong Leav Sangkat, Prey Veng Municipality, Prey Veng Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Svay Rieng Provincial Branch",
+            category = Category.BANK,
+            position = Position(11.093722, 105.795130),
+            phone = "(+855) 96 845 7777",
+            address = "National Road No. 1, Kien Sang Village, Svay Rieng Sangkat, Svay Rieng Municipality, Svay Rieng Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Krong Bavet Branch - Sangkat Chrak Mtes",
+            category = Category.BANK,
+            position = Position(11.038932, 106.014319),
+            phone = "(+855) 96 807 9777",
+            address = "National Road No. 1, Tuol Ampil Village, Sangkat Chrak Mtes, Bavet Municipality, Svay Rieng Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Krong Bavet Branch",
+            category = Category.BANK,
+            position = Position(11.038311, 106.012141),
+            phone = "(+855) 96 304 7777",
+            address = "Bavet Kandal Village ,Bavet Sangkat,Bavet Municipality, Svay Rieng Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Kratie Provincial Branch",
+            category = Category.BANK,
+            position = Position(12.492904, 106.016979),
+            phone = "(+855) 96 854 7777",
+            address = "Preah Sihanouk Street, Phsar Veaeng Village, Kracheh Sangkat,Kracheh Municipality, Kratie Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Ratanak Kiri Provincial Branch",
+            category = Category.BANK,
+            position = Position(13.740640, 106.992775),
+            phone = "(+855) 96 855 7777",
+            address = "National Road No. 78, Phnom Svay Village, Boeng Kansaeng Sangkat, Ban Lung Municipality, Ratanakiri Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Koh Kong Provincial Branch",
+            category = Category.BANK,
+            position = Position(11.615376, 102.984587),
+            phone = "(+855) 96 871 7777",
+            address = "National Road No. 48, Phum Ti Muoy, Smach Mean Chey Sangkat, Khemara Phoumin Municipality, Koh Kong Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Pailin Provincial Branch",
+            category = Category.BANK,
+            position = Position(12.858804, 102.603091),
+            phone = "(+855) 96 869 7777",
+            address = "Ou Ta Puk Leu Village, Pailin Sangkat, Pailin Municipality, Pailin Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Stung Treng Provincial Branch",
+            category = Category.BANK,
+            position = Position(13.521203, 105.966011),
+            phone = "(+855) 96 866 7777",
+            address = "Reacheanukoul Village, Stung Treng Sangkat, Stung Treng Municipality, Stung Treng Province."
+        ),
+        Place(
+            name = "Sathapana Bank. Oddar Meanchey Provincial Branch",
+            category = Category.BANK,
+            position = Position(14.178147, 103.506667),
+            phone = "(+855) 96 870 7777",
+            address = "Samraong Village, Samraong Sangkat, Samraong Municipality, Oddar Meanchey Province."
+        ),
+        Place(
+            name = "Acleda Bank. Head Office",
+            category = Category.BANK,
+            position = Position(11.577880, 104.916854),
+            phone = "015 900 547",
+            address = "#61,Preah Monivong Blvd., Sangkat Srah Chork Khan Daun Penh, Phno Penh"
+        ),
+        Place(
+            name = "Wing. Head Office",
+            category = Category.BANK,
+            position = Position(11.545771, 104.922021),
+            phone = "",
+            address = "No 721, Preah Monivong blvd, Phnom Penh"
+        ),
+        Place(
+            name = "Chip Mong Commercial Bank. Mao Tse Tung",
+            category = Category.BANK,
+            position = Position(11.5440607, 104.9167246),
+            phone = "081 811 911",
+            address = "#137, Mao Tse Tung Blvd., Sangkat Boeung Keng Kang 3, Khan Chamkarmon, Phnom Penh."
+        ),
+        Place(
+            name = ". Preah Norodom",
+            category = Category.BANK,
+            position = Position(11.5616507, 104.9264938),
+            phone = "081 811 911",
+            address = "#99, Preah Norodom Blvd. Corner Samdech Pan Ave (Street 214), Sangkat Boeung Reang, Khan Doun Penh, Phnom Penh."
+        ),
+        Place(
+            name = ". Tuol Kouk ",
+            category = Category.BANK,
+            position = Position(11.5909478, 104.8921245),
+            phone = "081 811 911",
+            address = "#16 & 18, Street 1928, Sangkat Tuol Sangkae 2, Khan Russey Keo, Phnom Penh."
+        ),
+        Place(
+            name = ". Boeng Salang",
+            category = Category.BANK,
+            position = Position(11.5494203, 104.8906749),
+            phone = "081 811 911",
+            address = "#19H, Street 271, Sangkat Boeng Salang, Khan Tuol Kouk, Phnom Penh."
+        ),
+        Place(
+            name = ". Phsar Deum Thkov",
+            category = Category.BANK,
+            position = Position(11.5402552, 104.908997),
+            phone = "081 811 911",
+            address = "#37, Street 271, Sangkat Boeng Tumpun 1, Khan Mean Chey, Phnom Penh."
+        ),
+        Place(
+            name = ". Noro Mall",
+            category = Category.BANK,
+            position = Position(11.5485433, 104.9272106),
+            phone = "081 811 911",
+            address = "Corner Street 360 & Street 370), Sangkat Boeng Keng Kang Ti Muoy, Khan Boeng Keng Kang, Phnom Penh."
+        ),
+        Place(
+            name = ". Bak Touk",
+            category = Category.BANK,
+            position = Position(11.5631253, 104.9121438),
+            phone = "081 811 911",
+            address = "#174, Czech Republic Blvd. (Street 168) Corner Street 164, Sangkat Veal Vong, Khan Prampir Meakkakra, Phnom Penh."
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(11.5487114, 104.9064933),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "#175ABCD, Mao Tse Toung Blvd. Sangkat Toul Svay Prey Ti Mouy, Khan Chamkar Mon, Phnom Penh."
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(11.5900833, 104.8955833),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "#211, st 598, Sangkat Tuol Sangker 2, Khan Russey Keo, Phnom Penh."
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(11.5344722, 104.827500),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "#2111, National Road 4, Phum Prey Pring Khang Cheung 1, Sangkat Chaom Chau 3, Khan Por Senchey, Phnom Penh."
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(13.1064722, 103.1998333),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "National Road 5, Prek Moha Tep, Sangkat Svay Por, Krong Battambang, Battambang Province."
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(13.3676389, 103.8497222),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "National Road 6, Phum Salakanseng, Sangkat Svay Dankum, Krong Siem Reap, Siem Reap. "
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(11.9879167, 105.4604167),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "National Road 7, Phum Ti 4, Sangkat Vealvong, krong kampong, Kampong Cham. "
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(10.622000, 103.5256667),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "Phum Muoy, Sangkat Pir, Krong Preah Sihanouk, Preah Sihanouk Province."
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(11.476500, 104.9476944),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "National Road No. 21, Thmei Village, Sangkat Ta Khmao, Krong Ta Khmao, Kandal Province."
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(11.6492778, 104.9142778),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "No. 8 & 10, National Road No. 6A, Khtor Village, Sangkat Preaek Lieb, Khan Chraoy Chongvar, Phnom Penh."
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(11.4893889, 105.3289167),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "Lekh Prampir Village, Sangkat Kampon Leav, Krong Prey Veng, Prey Veng Province."
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(11.089442, 105.798719),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "Phum Suon Thmei, Sangkat Prei Chlak, Svay Rieng City, Svay Rieng "
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(12.4846667, 106.0195278),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "Doun Chroam Village, Sangkat Kracheh, Krong Kracheh Kratie, Province"
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(13.7393056, 106.9880278),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "Chey Chumnas Village, Sangkat Labansiek, Krong Ban Lung, Ratanak Kiri Province"
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(11.561010, 104.911415),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "Bldg S2-01 The Olympia City, preah Monireth Blvd. (217) Tep Phan Street (182), Sangkat Veal Vong, Khan Prampir Meakkakra, Phnom Penh"
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(11.5338333, 104.9296111),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "Lot No. 271, Norodom Blvd., Sangkat Tonle Basak, Khan Chamkar Mon, Phnom Penh"
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(11.460357, 104.512173),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "Lot No 884, Peanicheakkam Village, Sangkat Rokar Thum, Krong Chbar Mon, Kampong Speu Province. "
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(10.985461, 104.780354),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "Lot No 792, Louri Village, Sangkat Roka Knong, Krong Doun Kaev, Takeo Province"
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(13.807278, 104.978684),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "Andoung Pou Village, Sangkat Kampong Pranak, Krong Preah Vihear, Preah Vihear Province "
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(12.243362, 104.670823),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "La Tuek Trei Village, Sangkat Kampong Chhnang, Krong Kampong Chhnang, Kampong Chhnang Provice. "
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(12.703348, 104.978684),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "Lot No 1200, Kampong Thum Village, Sangkat Kampong Roteh, Krong Stueng Saen, Kampong Thom Province. "
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(11.531029, 104.944923),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "Bldg 218-218A, National Road No. 1, Phum Russey Sros, Sangkat Nirouth, Sangkat Nirouth Branch, Phnom Penh"
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(13.6541222, 102.5621316),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "Phum Baliley 1, Sangkat Poi Pet, Krong Poi Pet, Banteay Meanchey Province"
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(12.536575, 103.910052),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "# 0532, National Road 5, Stha Ni Village,Sangkat Svay Aat, Pursat Town, Pursat Province"
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(13.590950, 102.973928),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "N°AIII 1188,National Road N°6A,Phum Muoy ,Sangkat Preah Ponlea , Krong Serei Saophoan, Banteay Meanchey Province"
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(11.533155, 104.913000),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "#363, St271, Group17, Village3, Commune Phsar Deum Thkov, District Chamka Morn, Phnom Penh"
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(13.530700, 105.970800),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "Phum Preaek, Sangkat Stung Treng, Krong Stung Treng, Stung Treng Province"
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(11.912351, 105.647375),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "Cheung Lang Village, Sangkat Suong, Krong Suong, Tboung Khmum Province"
+        ),
+        Place(
+            name = "Prince Bank. Branch",
+            category = Category.BANK,
+            position = Position(10.608400, 104.180800),
+            phone = "1800-20-8888 ( Free of Charge )",
+            address = "Muoy Usaphea Village,Sangkat Kampong Kandal,Krong Kampot,Kampot Province"
+        ),
+        Place(
+            name = "Prasac MFI. Head Office",
+            category = Category.BANK,
+            position = Position(11.539201, 104.910071),
+            phone = "081 421 572 / 016 444 141",
+            address = "Building № 212, Street 271, Tuol Tumpung 2, Chamkarmon, Phnom Penh, Cambodia"
+        ),
+        Place(
+            name = "Prasac MFI. Phsar Dei Hoy",
+            category = Category.BANK,
+            position = Position(11.572344, 104.870346),
+            phone = "016 444 004",
+            address = "St. 1019, Roung Chak Village, Sangkat Phnom Penh Thmei, Khan Sensok, Phnom Penh"
+        ),
+        Place(
+            name = "Prasac MFI. Vimean Ekareach",
+            category = Category.BANK,
+            position = Position(11.555116, 104.928799),
+            phone = "015 222 052",
+            address = "House № 124, Preah Norodom Bvld., Phum 4, Sangkat Tonle Basak, Khan Chamkar Mon , Phnom Penh"
+        ),
+        Place(
+            name = "Prasac MFI. Tuol Kouk",
+            category = Category.BANK,
+            position = Position(11.584253, 104.900544),
+            phone = "015 698 511",
+            address = "House № 150 G2, St. 289, Group 13, Sangkat Boeng Kak I, Khan Tuol Kouk, Phnom Penh"
+        ),
+        Place(
+            name = "Prasac MFI. Bak Touk",
+            category = Category.BANK,
+            position = Position(11.564369, 104.911493),
+            phone = "093 944 013",
+            address = "St. 169"
+        ),
+        Place(
+            name = "Prasac MFI. Chaom Chau",
+            category = Category.BANK,
+            position = Position(11.533644, 104.822420),
+            phone = "016 690 933",
+            address = "House № 94, National Road 4, Praek Pring khangcherng, Sangkat Chaom Chao, Khan Posen Chey, Phnom Penh"
+        ),
+        Place(
+            name = "Prasac MFI. Stueng Mean chey",
+            category = Category.BANK,
+            position = Position(11.534404, 104.881031),
+            phone = "016 690 934",
+            address = "House № 101, St. Veng Sreng, Domnak Thum 3 Village , Sangkat Stueng Mean Chey, Khan Meanchey, Phnom Penh"
+        ),
+        Place(
+            name = "Prasac MFI. Phsar Daeum Thkov",
+            category = Category.BANK,
+            position = Position(11.527028, 104.917109),
+            phone = "015 697 529",
+            address = "St. 271, Phum Khbal Tumnorp, Phnom Penh"
+        ),
+        Place(
+            name = "Prasac MFI. Phsar Heng Ly",
+            category = Category.BANK,
+            position = Position(11.554827, 104.887125),
+            phone = "016 201 753",
+            address = "House № 1E-4D, St. 271, Phum 11, Sangkat Tuek L'ak 3, Khan Tuol Kouk, Phnom Penh"
+        ),
+        Place(
+            name = "Prasac MFI. Phsar Kandal",
+            category = Category.BANK,
+            position = Position(11.571558, 104.927651),
+            phone = " 016 301 071",
+            address = "House № 102, St. 13, Phum 14, Sangkat Phsar Kandal 1, Khan Daun Penh, Phnom Penh"
+        ),
+        Place(
+            name = "Prasac MFI. Monivong",
+            category = Category.BANK,
+            position = Position(11.548335, 104.921515),
+            phone = "016 201 436",
+            address = "Preah Monivong Blvd., Phum Ti 7, Sangkat Boeng Keng Kang 3, Khan Boeng Keng Kang , Phnom Penh"
+        ),
+        Place(
+            name = "Prasac MFI. Chbar Ampov",
+            category = Category.BANK,
+            position = Position(11.532014, 104.938418),
+            phone = "016 985 877",
+            address = "National Road 1, Daem Slaeng Village, Sangkat Chbar Ampov 2, Khan Chbar Ampov, Phnom Penh"
+        ),
+        Place(
+            name = "Prasac MFI. Praek Pnov",
+            category = Category.BANK,
+            position = Position(11.657695, 104.864138),
+            phone = "016 201 731",
+            address = "Kandal Village, Sangkat Prek Pnov, Khan Prek Pnov, Phnom Penh"
+        ),
+        Place(
+            name = "Prasac MFI. Dangkao",
+            category = Category.BANK,
+            position = Position(11.519047, 104.889339),
+            phone = "015 222 054",
+            address = "House № A09, St. 217 (Chamkar Doung St.), Sombour Village, Sangkat Dangkao, Khan Dangkao, Phnom Penh"
+        ),
+        Place(
+            name = "Prasac MFI. Pochentong",
+            category = Category.BANK,
+            position = Position(11.563251, 104.862433),
+            phone = "016 200 695",
+            address = "House № 28 Eo, St. 110 (Russian Federation Blvd.), Paprak Choeung Village, Sangkat Kakab, Khan Posen Chey, Phnom Penh"
+        ),
+        Place(
+            name = "Prasac MFI. Preaek Lieb",
+            category = Category.BANK,
+            position = Position(11.592101, 104.929479),
+            phone = "015 222 052",
+            address = "House № 206 G, National Road 6, Phum 3, Sangkat Chraoy Chongvar, Khan Chraoy Chongvar, Phnom Penh"
+        ),
+        Place(
+            name = "Prasac MFI. Boeng Keng Kang Ti Bei",
+            category = Category.BANK,
+            position = Position(11.547987, 104.913996),
+            phone = "016 444 068",
+            address = "House № 436 CD, Corner St. 163 & St. 360, Phum 2, Sangkat Boeng Keng Kang 3, Khan Boeng Keng Kang, Phnom Penh"
+        ),
+        Place(
+            name = "Prasac MFI. Phsar Thmey",
+            category = Category.BANK,
+            position = Position(11.569524, 104.920121),
+            phone = "081 444 988",
+            address = "House № 30Eo, E1, E2, E3, Street 67 corner 128, Sankat Phsar Thmei 2 Khan Doun Penh, Phhom Penh"
+        ),
+        Place(
+            name = "Prasac MFI. Boeng Trabaek",
+            category = Category.BANK,
+            position = Position(11.538023, 104.923538),
+            phone = "016 657 687",
+            address = "House 460A-460B, Monivong Blvd, Street 466, Sangkat Tonle Basak, Khan Chamkar Mon, Phnom Penh"
+        ),
+        Place(
+            name = "Prasac MFI. Bokor",
+            category = Category.BANK,
+            position = Position(11.543860, 104.920145),
+            phone = "023 999 911",
+            address = "House, 40ABC, Mao Tse Toung Blvd, Sangkat Boeng Trabaek, Khan Chamkar Mon, Phnom Penh"
+        ),
+        Place(
+            name = "Prasac MFI. Kandal",
+            category = Category.BANK,
+            position = Position(11.476552, 104.941719),
+            phone = "015 697 408",
+            address = "House № 234, National Road 2, Krapeu Ha Village, Prek Russie Commune , Krong Takhmao, Kandal Province"
+        ),
+        Place(
+            name = "Prasac MFI. Banteay Meanchey",
+            category = Category.BANK,
+            position = Position(13.587298, 102.969965),
+            phone = "015 698 130",
+            address = "National Raod 5, Korothan Village, Sangkat Ou Ambel, Krong Sereisophon, Banteay Meanchey Province"
+        ),
+        Place(
+            name = "Prasac MFI. Battambang",
+            category = Category.BANK,
+            position = Position(13.102589, 103.207057),
+            phone = "015 698 488",
+            address = "National Road 5, Rumcheak 5 Village, Sangkat Ratanak, Krong Battambang, Battambang Province"
+        ),
+        Place(
+            name = "Prasac MFI. Kampong Cham",
+            category = Category.BANK,
+            position = Position(11.988508, 105.459928),
+            phone = "015 697 319",
+            address = "National Raod 7, 4 V,Sangkat Veal Vong,Krong Kampong Cham, Kampong Cham Province"
+        ),
+        Place(
+            name = "Prasac MFI. Kampong Chhnang",
+            category = Category.BANK,
+            position = Position(12.240623, 104.670466),
+            phone = "015 697 924",
+            address = " National Road 5, Srae Pring Village, Sangkat Kampong Chhnang, Krong Kampong Chhnang, Kampong Chhnang Province"
+        ),
+        Place(
+            name = "Prasac MFI. Kampong Speu",
+            category = Category.BANK,
+            position = Position(11.462111, 104.518698),
+            phone = "015 697 058",
+            address = " National Road 4, Krang Pol Tep Village, Sangkat Kampong Roka Thom, Krong Chbar Morn, Kompong Speu Province"
+        ),
+        Place(
+            name = "Prasac MFI. Kampong Thom",
+            category = Category.BANK,
+            position = Position(12.7042188, 104.8907795),
+            phone = "015 667 007",
+            address = "National Road 6A, Kampong Thom Village, Kampong Rotes Commune, Steung Sen District, Kampong Thom Province"
+        ),
+        Place(
+            name = "Prasac MFI. Kampot",
+            category = Category.BANK,
+            position = Position(10.6190898, 104.1830348),
+            phone = "015 697 587",
+            address = "House № 21, National Road 3, Kampong bay Tbong Village, Sangkat Kampongbay, Krong Kampot, Kampot Province"
+        ),
+        Place(
+            name = "Prasac MFI. Angk Snuol",
+            category = Category.BANK,
+            position = Position(11.5092036, 104.7418472),
+            phone = "015 698 604",
+            address = "National Road 4, Svay Chrum Village, Khum Baek Chan , Angk Snoul District, Kandal Province"
+        ),
+        Place(
+            name = "Prasac MFI. Kien Svay",
+            category = Category.BANK,
+            position = Position(11.5112203, 105.0478315),
+            phone = "015 697 485",
+            address = "National Road 1, Korki Village, Korki Commune, Kien Svay District, Kandal Province"
+        ),
+        Place(
+            name = "Prasac MFI. Mukh Kampul",
+            category = Category.BANK,
+            position = Position(11.7317337, 104.97209),
+            phone = "015 698 620",
+            address = " Kraom Village , Prek Anhchanh Commune , Mukh Kampol District, Kandal Province"
+        ),
+        Place(
+            name = "Prasac MFI. Koh Kong",
+            category = Category.BANK,
+            position = Position(11.6154339, 102.9844243),
+            phone = "015 698 771",
+            address = "National Road 48, Phum 1, Sangkat Smach Meanchey, Krong Khemarak Phoumint, Koh Kong Province"
+        ),
+        Place(
+            name = "Prasac MFI. Kracheh",
+            category = Category.BANK,
+            position = Position(12.4875978, 106.0186673),
+            phone = "015 698 664",
+            address = "Doun Chroam Village, Kratie Commune , Krong Kratie, Kratie Province"
+        ),
+        Place(
+            name = "Prasac MFI. Mondul Kiri",
+            category = Category.BANK,
+            position = Position(12.4593797, 107.186611),
+            phone = " 015 698 087",
+            address = "Kandal Village, Sangkat Spean Mean Chey, Krong Saen Monourom, Mondul Kiri Province"
+        ),
+        Place(
+            name = "Prasac MFI. Preah Vihear",
+            category = Category.BANK,
+            position = Position(13.8018379, 104.9791021),
+            phone = " 015 698 864",
+            address = "National Road 62, Andoung Pou Village, Sangkat Kompong Pranak, Krong Preah Vihear, Preah Vihear Province"
+        ),
+        Place(
+            name = "Prasac MFI. Prey Veng",
+            category = Category.BANK,
+            position = Position(11.4862321, 105.3255333),
+            phone = "015 697 115",
+            address = "National Road 11, Phum 4, Sankat Kampong Leav, Prey Veng District, Prey Veng Province"
+        ),
+        Place(
+            name = "Prasac MFI. Pursat",
+            category = Category.BANK,
+            position = Position(12.5366692, 103.9108875),
+            phone = "015 698 809",
+            address = " National Road 5, Ra Village, Sankat Phpas Prey, Krong Pursat, Pursat Province"
+        ),
+        Place(
+            name = "Prasac MFI. Ratanak Kiri",
+            category = Category.BANK,
+            position = Position(13.738019, 106.982013),
+            phone = "015 698 660",
+            address = "National Road 78, Phum 7Meakara , Sangkat Labansiek, Krong Banlung, Ratanak Kiri Province"
+        ),
+        Place(
+            name = "Prasac MFI. Siem Reap",
+            category = Category.BANK,
+            position = Position(13.3602175, 103.8642184),
+            phone = "015 698 939",
+            address = "National Road 6, Banteay Chas Village, Sangkat Slor Kram, Krong Siem Reap, Siem Reap Province"
+        ),
+        Place(
+            name = "Prasac MFI. Svay Dankum",
+            category = Category.BANK,
+            position = Position(13.3460703, 103.8483592),
+            phone = "016 201 651",
+            address = "National Raod 6, Vihear Chen Village, Svay Dankum Commune, Siem Reap District, Siem Reap Province"
+        ),
+        Place(
+            name = "Prasac MFI. Preah Sihanouk",
+            category = Category.BANK,
+            position = Position(10.6227879, 103.5251106),
+            phone = "015 698 971",
+            address = "St. Ekareach, Phum 1, Sangkat 2, Krong Preah Sihanouk, Preah Sihanouk Province"
+        ),
+        Place(
+            name = "Prasac MFI. Stung Treng",
+            category = Category.BANK,
+            position = Position(13.5218648, 105.9669708),
+            phone = "015 698 653",
+            address = "St. 63, Trapeang Pring Village, Sangkat Stueng Treng, Stueng Treng District, Stueng Treng Province"
+        ),
+        Place(
+            name = "Prasac MFI. Svay Rieng",
+            category = Category.BANK,
+            position = Position(11.090373, 105.7949341),
+            phone = "016 200 540",
+            address = " National Road 1, Kien Sang Village, Sangkat Svay Rieng, Krong Svay Rieng, Svay Rieng Province"
+        ),
+        Place(
+            name = "Prasac MFI. Takeo",
+            category = Category.BANK,
+            position = Position(10.9793008, 104.7776557),
+            phone = "015 697",
+            address = "House № 182, National Road 2, Thnal Baeck Village, Sangkat Rokakav, Krong Daunkeo, Takeo Province"
+        ),
+        Place(
+            name = "Prasac MFI. Otdar Meanchey",
+            category = Category.BANK,
+            position = Position(14.1705832, 103.5121246),
+            phone = "015 698 964",
+            address = "House № 68, Phum Donkeo, Sangkat Samrong, Krong Samrong, Otdor Meanchey Province"
+        ),
+        Place(
+            name = "Prasac MFI. Pailin",
+            category = Category.BANK,
+            position = Position(12.8557462, 102.6076289),
+            phone = "015 698 623",
+            address = "Podin Neav Village, Sangkat Pailin, Krong Pailin, Pailin Province"
+        ),
+        Place(
+            name = "Prasac MFI. Steung Sangkae",
+            category = Category.BANK,
+            position = Position(13.0961847, 103.1984348),
+            phone = "015 444 217",
+            address = "St.01 No.458-459 Kampong Krabei Village, Sangkat Svay Pao, Krong Battambang, Battambang Province"
+        ),
+        Place(
+            name = "Prasac MFI. Suong",
+            category = Category.BANK,
+            position = Position(11.9113595, 105.6507608),
+            phone = "015 698 030",
+            address = "National Raod 7, Phum Cheung Lang, Suong, Krong Suong, Tboung Khmum Province"
+        ),
+        Place(
+            name = "Prasac MFI. Svay Karvan",
+            category = Category.BANK,
+            position = Position(11.4551847, 104.5208373),
+            phone = "069 379 742",
+            address = "Tuol Kouk Village, Sangkat Svay Kravan, Krong Chbar Mon, Kampong Speu Province"
+        ),
+        Place(
+            name = "Prasac MFI. Phsar Deam Ampil",
+            category = Category.BANK,
+            position = Position(11.3679535, 104.8666203),
+            phone = "016 201 627",
+            address = "National Road #2, Pralay Village, Boeng Khyang Commune Kandal Stueng District, Kandal Province"
+        ),
+        Place(
+            name = "Prasac MFI. Sa Ang",
+            category = Category.BANK,
+            position = Position(11.3606316, 105.0053922),
+            phone = "015",
+            address = " National Road 21, Prek Run Village, Prek Koy Commune , Sa Ang District, Kandal Province"
+        ),
+        Place(
+            name = "Prasac MFI. Kandal Stoeng",
+            category = Category.BANK,
+            position = Position(11.4292795, 104.8114819),
+            phone = "015",
+            address = "National Road 51, Svay Ming Village, Bakou Commune , Kandal Stueng District, Kandal Province"
+        ),
+        Place(
+            name = "FTB. Head Office Branch",
+            category = Category.BANK,
+            position = Position(11.5717833, 104.9214524),
+            phone = "023 862 111",
+            address = "No. 4A, Corner of street 114 & 53, R.V Kramoun Sar, Sangkat Phsar Tmey I, Kh. Daun Penh, Phnom Penh."
+        ),
+        Place(
+            name = "FTB. Olympic Branch",
+            category = Category.BANK,
+            position = Position(11.5530279, 104.911393),
+            phone = "023 862 111",
+            address = "No.392E0, St. 193, Olympic, Chamkarmorn, Phnom Penh"
+        ),
+        Place(
+            name = "FTB. Samdach Pan Branch",
+            category = Category.BANK,
+            position = Position(11.5604171, 104.9225068),
+            phone = "023 862 111",
+            address = "No. 49, St. 214 corner St.63, Boeung Raing, Daun Penh, Phnom Penh"
+        ),
+        Place(
+            name = "FTB. Kompong Cham Branch",
+            category = Category.BANK,
+            position = Position(11.9909092, 105.4559007),
+            phone = "023 862 111",
+            address = "Kompong Cham"
+        ),
+        Place(
+            name = "FTB. Bak Touk Branch",
+            category = Category.BANK,
+            position = Position(11.5634122, 104.9116834),
+            phone = "023 862 111",
+            address = "No. 33C-D, Street 169, Sangkat Veal Vong, Khan 7Makara, Phnom Penh"
+        ),
+        Place(
+            name = "FTB. Siem Reap Branch",
+            category = Category.BANK,
+            position = Position(13.3539276, 103.8547326),
+            phone = "023 862 111",
+            address = "Siem Reap"
+        ),
+        Place(
+            name = "FTB. IFL Branch",
+            category = Category.BANK,
+            position = Position(11.5685986, 104.8920964),
+            phone = "023 862 111",
+            address = "Russian Federation Blvd, Phnom Penh Cambodia"
+        ),
+        Place(
+            name = "FTB. Battambang Branch",
+            category = Category.BANK,
+            position = Position(13.1047209, 103.2002136),
+            phone = "023 862 111",
+            address = "Battambang"
+        ),
+        Place(
+            name = "FTB. Orkide Branch Street 2004",
+            category = Category.BANK,
+            position = Position(11.5520556, 104.8735891),
+            phone = "023 862 111",
+            address = "Borey Orkide #51-53-55, Street 2004, Phum Okide, Sangkat Ou Baek K'am, Khan Sen Sok, Phnom Penh"
+        ),
+        Place(
+            name = "FTB. Sihanouk Psar Leur Branch",
+            category = Category.BANK,
+            position = Position(10.6245556, 103.5235613),
+            phone = "023 862 111",
+            address = "Sihanouk Ville"
+        ),
+        Place(
+            name = "FTB. Phnom Penh Port Branch",
+            category = Category.BANK,
+            position = Position(11.5835307, 104.9206263),
+            phone = "023 862 111",
+            address = "No. 649, St. Preah Sisowat, Phnom Penh Port, Phnom Penh"
+        ),
+        Place(
+            name = "FTB. Sihanouk Port Branch",
+            category = Category.BANK,
+            position = Position(10.645, 103.5090891),
+            phone = "023 862 111",
+            address = "Sihanouk Ville"
+        ),
+        Place(
+            name = "FTB. Ta Khmao Branch ",
+            category = Category.BANK,
+            position = Position(11.4814288, 104.9450693),
+            phone = "023 862 111",
+            address = "National Road 2, St. 204, Phum Takhmao, Sangkat Takhmao, Krong Takhmao, Kandal province"
+        ),
+        Place(
+            name = "FTB. Toul Kork Branch",
+            category = Category.BANK,
+            position = Position(11.583083, 104.8983783),
+            phone = "023 862 111",
+            address = "No. 80, St. 315, TK Avenue Center, Sangkat Boeung Kok 1, Khan Toul Kok, Phnom Penh"
+        ),
+        Place(
+            name = "FTB. Psar Thom Tmey Branch",
+            category = Category.BANK,
+            position = Position(11.5697988, 104.9207512),
+            phone = "023 862 111",
+            address = "57 Prey Nokor St. (126), Phnom Penh, Cambodia"
+        )
+        // endregion
+    )
