+++ library/src/main/java/jp/co/soramitsu/map/data/PlacesRepository.kt
@@ -1,0 +1,7 @@
+package jp.co.soramitsu.map.data
+
+import jp.co.soramitsu.map.model.Place
+
+interface PlacesRepository {
+    fun getAllPlaces(): List<Place>
+}
