+++ library/src/main/java/jp/co/soramitsu/map/model/Author.kt
@@ -1,0 +1,6 @@
+package jp.co.soramitsu.map.model
+
+class Author(
+    val name: String,
+    val avatarUrl: String = ""
+)
