+++ library/src/main/java/jp/co/soramitsu/map/model/Category.kt
@@ -1,0 +1,12 @@
+package jp.co.soramitsu.map.model
+
+enum class Category {
+    BANK,
+    SERVICES,
+    FOOD,
+    SUPERMARKETS,
+    PHARMACY,
+    ENTERTAINMENT,
+    EDUCATION,
+    OTHER
+}
