+++ library/src/main/java/jp/co/soramitsu/map/model/Position.kt
@@ -1,0 +1,6 @@
+package jp.co.soramitsu.map.model
+
+class Position(
+    val latitude: Double,
+    val longitude: Double
+)
