+++ library/src/main/java/jp/co/soramitsu/map/model/Review.kt
@@ -1,0 +1,8 @@
+package jp.co.soramitsu.map.model
+
+data class Review(
+    val author: Author,
+    val rating: Float,
+    val date: Long,
+    val text: String
+)
