+++ library/src/main/java/jp/co/soramitsu/map/model/Time.kt
@@ -1,0 +1,6 @@
+package jp.co.soramitsu.map.model
+
+data class Time(
+    val hour: Int,
+    val minute: Int
+)
