+++ library/src/main/java/jp/co/soramitsu/map/presentation/GoogleMapActivity.kt
@@ -1,0 +1,13 @@
+package jp.co.soramitsu.map.presentation
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import jp.co.soramitsu.map.R
+
+class GoogleMapActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_google_map)
+    }
+}
