+++ library/src/main/res/layout/place_bottom_sheet.xml
@@ -23,7 +23,7 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <jp.co.soramitsu.feature.maps.impl.presentation.places.PlaceView
+    <jp.co.soramitsu.map.presentation.places.PlaceView
         android:id="@+id/placeListItem"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
